# docs/swagger.yaml
basePath: /api/v1
definitions:
  devops-cd_internal_dto.AppEnvConfigResponse:
    properties:
      app_id:
        type: integer
      cluster:
        type: string
      config_data:
        type: string
      created_at:
        type: string
      deployment_name_override:
        type: string
      env:
        type: string
      id:
        type: integer
      replicas:
        type: integer
      status:
        type: integer
      updated_at:
        type: string
    type: object
...
  devops-cd_internal_dto.ApplicationDependenciesResponse:
    properties:
      app_id:
        type: integer
      app_name:
        type: string
      dependencies:
        items:
          type: integer
        type: array
      updated_at:
        type: string
    type: object
  devops-cd_internal_dto.ApplicationResponse:
    properties:
      app_type:
        type: string
      created_at:
        type: string
      default_depends_on:
        items:
          type: integer
        type: array
      deployed_tag:
        description: 当前部署的镜像标签
        type: string
      description:
        type: string
      env_clusters:
        additionalProperties:
          items:
            type: string
          type: array
        description: 环境集群配置，从 app_env_configs 表查询得出
        type: object
      id:
        type: integer
      name:
        type: string
      project_id:
        description: 关联的项目ID
        type: integer
      project_name:
        description: 项目名称
        type: string
      repo_id:
        type: integer
      repo_name:
        description: Repository的namespace/name
        type: string
      status:
        type: integer
      team_id:
        type: integer
      team_name:
        description: 团队名称
        type: string
      updated_at:
        type: string
    type: object
  devops-cd_internal_dto.BatchCreateAppEnvConfigsRequest:
    properties:
      app_id:
        type: integer
      configs:
        items:
          $ref: '#/definitions/devops-cd_internal_dto.CreateAppEnvConfigItemRequest'
        minItems: 1
        type: array
    required:
    - app_id
    - configs
    type: object
...
  devops-cd_internal_dto.CreateAppEnvConfigItemRequest:
    properties:
      cluster:
        maxLength: 50
        type: string
      config_data:
        type: string
      deployment_name_override:
        maxLength: 63
        type: string
      env:
        enum:
        - pre
        - prod
        - dev
        - test
        - uat
        type: string
      replicas:
        maximum: 100
        minimum: 1
        type: integer
    required:
    - cluster
    - env
    - replicas
    type: object
  devops-cd_internal_dto.CreateAppEnvConfigRequest:
    properties:
      app_id:
        type: integer
      cluster:
        maxLength: 50
        type: string
      config_data:
        description: JSON字符串
        type: string
      deployment_name_override:
        maxLength: 63
        type: string
      env:
        enum:
        - pre
        - prod
        - dev
        - test
        - uat
        type: string
      replicas:
        maximum: 100
        minimum: 1
        type: integer
    required:
    - app_id
    - cluster
    - env
    - replicas
    type: object
  devops-cd_internal_dto.CreateApplicationRequest:
    properties:
      app_type:
        enum:
        - static
        - node
        - java
        - go
        - py
        type: string
      description:
        type: string
      display_name:
        type: string
      env_clusters:
        additionalProperties:
          items:
            type: string
          type: array
        description: 环境集群配置，用于初始化 app_env_configs 表
        type: object
      name:
        maxLength: 100
        type: string
      project_id:
        description: 可选：如果不指定则从 Repository 继承
        type: integer
      repo_id:
        type: integer
      team_id:
        type: integer
    required:
    - app_type
    - name
    - repo_id
    type: object
...
  devops-cd_internal_dto.DeleteApplicationRequest:
    properties:
      id:
        description: 必填：应用ID
        type: integer
    required:
    - id
    type: object
...
  devops-cd_internal_dto.PageResponse:
    properties:
      items: {}
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
...
  devops-cd_internal_dto.UpdateAppDependenciesRequest:
    properties:
      dependencies:
        items:
          type: integer
        type: array
    type: object
  devops-cd_internal_dto.UpdateAppEnvConfigRequest:
    properties:
      cluster:
        maxLength: 50
        type: string
      config_data:
        type: string
      deployment_name_override:
        maxLength: 63
        type: string
      id:
        type: integer
      replicas:
        maximum: 100
        minimum: 1
        type: integer
      status:
        enum:
        - 0
        - 1
        type: integer
    required:
    - id
    type: object
  devops-cd_internal_dto.UpdateApplicationRequest:
    properties:
      app_type:
        enum:
        - static
        - node
        - java
        - go
        - py
        type: string
      deployed_tag:
        description: 当前部署的镜像标签
        type: string
      description:
        type: string
      display_name:
        type: string
      env_clusters:
        additionalProperties:
          items:
            type: string
          type: array
        description: 环境集群配置，用于同步更新 app_env_configs 表
        type: object
      id:
        description: 必填：应用ID
        type: integer
      name:
        maxLength: 100
        type: string
      status:
        enum:
        - 0
        - 1
        type: integer
      team_id:
        type: integer
    required:
    - id
    type: object
...
  devops-cd_pkg_utils.Response:
    properties:
      code:
        type: integer
      data: {}
      detail:
        description: 详细错误信息（可选）
        type: string
      message:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: |-
    DevOps 持续交付平台 API 文档
    提供代码库管理、应用管理、批次管理、部署管理等功能
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: DevOps CD API
  version: "1.0"
paths:
  /api/v1/app-env-configs:
    get:
      consumes:
      - application/json
      parameters:
      - description: 应用ID
        format: int64
        in: query
        name: app_id
        required: true
        type: integer
      - description: 环境名称(pre/prod/dev/test/uat)
        in: query
        name: env
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/devops-cd_pkg_utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/devops-cd_internal_dto.AppEnvConfigResponse'
                  type: array
              type: object
      summary: 查询应用环境配置列表
      tags:
      - AppEnvConfig
    post:
      consumes:
      - application/json
      parameters:
      - description: 创建请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/devops-cd_internal_dto.CreateAppEnvConfigRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/devops-cd_pkg_utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/devops-cd_internal_dto.AppEnvConfigResponse'
              type: object
      summary: 创建应用环境配置
      tags:
      - AppEnvConfig
  /api/v1/app-env-configs/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 配置ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/devops-cd_pkg_utils.Response'
      summary: 删除应用环境配置
      tags:
      - AppEnvConfig
    get:
      consumes:
      - application/json
      parameters:
      - description: 配置ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/devops-cd_pkg_utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/devops-cd_internal_dto.AppEnvConfigResponse'
              type: object
      summary: 获取应用环境配置详情
      tags:
      - AppEnvConfig
    put:
      consumes:
      - application/json
      parameters:
      - description: 配置ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      - description: 更新请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/devops-cd_internal_dto.UpdateAppEnvConfigRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/devops-cd_pkg_utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/devops-cd_internal_dto.AppEnvConfigResponse'
              type: object
      summary: 更新应用环境配置
      tags:
      - AppEnvConfig
  /api/v1/app-env-configs/batch:
    post:
      consumes:
      - application/json
      parameters:
      - description: 批量创建请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/devops-cd_internal_dto.BatchCreateAppEnvConfigsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/devops-cd_pkg_utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/devops-cd_internal_dto.AppEnvConfigResponse'
                  type: array
              type: object
      summary: 批量创建应用环境配置
      tags:
      - AppEnvConfig
  /api/v1/application:
    get:
      consumes:
      - application/json
      parameters:
      - description: 应用ID
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/devops-cd_pkg_utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/devops-cd_internal_dto.ApplicationResponse'
              type: object
      summary: 获取应用详情
      tags:
      - Application
    put:
      consumes:
      - application/json
      parameters:
      - description: 更新应用请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/devops-cd_internal_dto.UpdateApplicationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/devops-cd_pkg_utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/devops-cd_internal_dto.ApplicationResponse'
              type: object
      summary: 更新应用
      tags:
      - Application
  /api/v1/application/{id}/dependencies:
    get:
      parameters:
      - description: 应用ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/devops-cd_pkg_utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/devops-cd_internal_dto.ApplicationDependenciesResponse'
              type: object
      summary: 获取应用默认依赖
      tags:
      - Application
    put:
      consumes:
      - application/json
      parameters:
      - description: 应用ID
        in: path
        name: id
        required: true
        type: integer
      - description: 更新依赖请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/devops-cd_internal_dto.UpdateAppDependenciesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/devops-cd_pkg_utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/devops-cd_internal_dto.ApplicationDependenciesResponse'
              type: object
      summary: 更新应用默认依赖
      tags:
      - Application
  /api/v1/application/builds:
    get:
      consumes:
      - application/json
      parameters:
      - description: 应用ID
        in: query
        name: id
        required: true
        type: integer
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/devops-cd_pkg_utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/devops-cd_internal_dto.PageResponse'
              type: object
      summary: 获取应用的构建历史
      tags:
      - Application
  /api/v1/application/delete:
    post:
      consumes:
      - application/json
      parameters:
      - description: 删除应用请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/devops-cd_internal_dto.DeleteApplicationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/devops-cd_pkg_utils.Response'
      summary: 删除应用
      tags:
      - Application
...
  /api/v1/application_builds:
    get:
      consumes:
      - application/json
      description: 搜索应用列表，支持模糊查询 app、repo、commit、tag 等字段，返回包含最新构建信息的应用列表，按构建时间倒序排序
      parameters:
      - description: 模糊查询关键字（支持查询 app、repo、commit、tag 等字段）
        in: query
        name: keyword
        type: string
      - description: 页码（默认1）
        in: query
        name: page
        type: integer
      - description: 每页数量（默认20，最大100）
        in: query
        name: page_size
        type: integer
      - description: 按代码库ID过滤
        in: query
        name: repo_id
        type: integer
      - description: 按团队ID过滤
        in: query
        name: team_id
        type: integer
      - description: 按应用类型过滤
        in: query
        name: app_type
        type: string
      - description: 按状态过滤（0/1）
        in: query
        name: status
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/devops-cd_pkg_utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/devops-cd_internal_dto.PageResponse'
              type: object
      summary: 搜索应用（包含构建信息）
      tags:
      - Application
  /api/v1/applications:
    get:
      consumes:
      - application/json
      parameters:
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量
        in: query
        name: page_size
        type: integer
      - description: 代码库ID
        in: query
        name: repo_id
        type: integer
      - description: 团队ID
        in: query
        name: team_id
        type: integer
      - description: 应用类型
        in: query
        name: app_type
        type: string
      - description: 关键字
        in: query
        name: keyword
        type: string
      - description: 状态
        in: query
        name: status
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/devops-cd_pkg_utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/devops-cd_internal_dto.PageResponse'
              type: object
      summary: 获取应用列表
      tags:
      - Application
    post:
      consumes:
      - application/json
      parameters:
      - description: 创建应用请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/devops-cd_internal_dto.CreateApplicationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/devops-cd_pkg_utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/devops-cd_internal_dto.ApplicationResponse'
              type: object
      summary: 创建应用
      tags:
      - Application
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"